add test and move to correct modules
