+++ src/lc3/bytes.rs
@@ -28,3 +28,17 @@ pub fn concatenate_bytes(bytes: &[u8]) -> Result<u16, VMError> {
     let res = u16::from_be_bytes([first_byte, second_byte]);
     Ok(res)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn sign_extend_mantains_sign() -> Result<(), VMError> {
+        let x = 0b11111;
+        let y = 0b01111;
+        assert_eq!(sign_extend(x, 5)?, 0xFFFF);
+        assert_eq!(sign_extend(y, 5)?, 0x000F);
+        Ok(())
+    }
+}
